+++ packages/polaris-viz/src/utilities/getAnimationDelayForItems.ts
@@ -1,0 +1,10 @@
+import {DEFAULT_ANIMATION_DELAY} from '../constants';
+
+export function getAnimationDelayForItems(count: number) {
+  const scalingFactor = count * 0.1;
+
+  return Math.min(
+    DEFAULT_ANIMATION_DELAY / scalingFactor,
+    DEFAULT_ANIMATION_DELAY,
+  );
+}